Add files via upload
V1 of the assembly showed up several errors. For example, if the arm was in a rotated position the flapper bar lever would disengage - so I have opted for a different mechanism similar to the mechanism in the robot this is all based on.

I have also rearranged the cogs so they are directly above one another rather than at an angle. This should make future calculations easier.